+++ src/dungeon/dungeon.h
@@ -6,7 +6,7 @@
 #include <glm/vec2.hpp>
 #include <cstdint>
 #include "sprite_renderer.h"
-#include "dungeon_tile.h"
+#include "dungeon/dungeon_tile.h"
 #include "text_rendering/text_renderer.h"
 #include "enemy.h"
 #include "input.h"
